requirements.txt
pandas==2.3.2
sqlalchemy==2.0.43
psycopg2-binary==2.9.11
streamlit==1.54.0
plotly==6.5.2
Faker==40.5.1
python-dotenv
